stripeapi_test/requirements.txt
django==4.2.8
flake8==6.1.0
isort==5.13.2
python-dotenv==1.0.0
djangorestframework==3.14.0
stripe==7.9.0